Our app as of July 22.
Toni and Trent are graphics and AR. Toni has created the alligator figure, various logos, the startup screen and the graphics for the java game. Trent designed the map which functions as our home page, and made the buttons that lead to separate activities. Miley started researching how to create the java game incorporating the graphics Toni made. Alyssa has designed the outline of the application in android studio. The app activities were created and linked to buttons on the layouts, and the buttons on the home page change color when clicked on, and are cleared back to normal when the user goes back to the startup screen.